Added one command to start the feeder for 5 seconds
